fix: 検索処理はDIしておく
