+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/data/api/GitHubApiClient.kt
@@ -1,0 +1,10 @@
+package jp.co.yumemi.android.code_check.data.api
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.android.Android
+
+object GitHubApiClient {
+    val client = HttpClient(Android) {
+        expectSuccess = true
+    }
+}
