+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/domain/GitHubRepository.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.code_check.domain
+
+import jp.co.yumemi.android.code_check.Item
+
+interface GitHubRepository {
+    suspend fun searchRepositories(query: String):List<Item>
+}
